requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.6
kombu==4.2.1
Markdown==2.6.11
PyJWT==1.6.4
pytz==2018.4
requests==2.18.4
urllib3==1.22
vine==1.1.4
